Move addNotification method to utils
