+++ plugins/tracker-resources/src/utils.ts
@@ -36,7 +36,11 @@ import {
   areDatesEqual,
   getMillisecondsInMonth,
   isWeekend,
-  MILLISECONDS_IN_WEEK
+  MILLISECONDS_IN_WEEK,
+  Notification,
+  NotificationPosition,
+  NotificationSeverity,
+  notificationsStore
 } from '@hcengineering/ui'
 import tracker from './plugin'
 import {
@@ -673,4 +677,25 @@ export function getWorkDayType (timestamp: number): WorkDaysType | undefined {
   } else if (areDatesEqual(date, previousWorkDate)) {
     return WorkDaysType.PREVIOUS
   }
+}
+
+export function addNotification (
+  id: string,
+  title: string,
+  subTitle: string,
+  component: AnyComponent | AnySvelteComponent,
+  params?: { [key: string]: any }
+): void {
+  const notification: Notification = {
+    id,
+    title,
+    subTitle,
+    severity: NotificationSeverity.Success,
+    position: NotificationPosition.BottomRight,
+    component,
+    closeTimeout: 10000,
+    params
+  }
+
+  notificationsStore.addNotification(notification)
 }
